crud complain

+++ src/services/complain/complain.service.ts
@@ -1,9 +1,11 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { AddComplainDto } from 'src/DTO/complains/addComplain';
 import { updateComplainDto } from 'src/DTO/complains/updateComplain';
+import { adminEntity } from 'src/entities/admin.entity';
 import { complainEntity } from 'src/entities/complain.entity';
-import { Repository } from 'typeorm';
+import { UserRoleEnum } from 'src/enums/user-role.enum';
+import { Any, Repository } from 'typeorm';
 
 @Injectable()
 export class ComplainService {
@@ -12,48 +14,105 @@ export class ComplainService {
     constructor(
         @InjectRepository(complainEntity)
         private readonly complainRepository: Repository<complainEntity>
-      )
+        
+        )
        {}
-       getComplain()
+
+
+       async getComplain(user): Promise<complainEntity[]>
         {
-          return this.complainRepository.find();
+          
+          if (user.role === UserRoleEnum.ADMINACHAT)
+             return await this.complainRepository.find();
+         
+         
+          throw new UnauthorizedException();
        }
-       async getById(id:string): Promise<complainEntity>
+
+
+       async getComplainClient(client): Promise<complainEntity[]>
+        {
+   
+           if (client.role === UserRoleEnum.CLIENT)
+               return await this.complainRepository.find({client});
+  
+  
+           throw new UnauthorizedException();
+        }
+
+       async getById(id:string, user): Promise<complainEntity>
        {
          const complain =  await this.complainRepository.findOne(id);
-         if (complain)
-           return complain;
-         else
-         throw new NotFoundException(`complain d'id ${id} n'existe pas`);
+         if (!complain)
+            throw new NotFoundException(`complain d'id ${id} n'existe pas`);
+         
+         if (user.role === UserRoleEnum.ADMINACHAT || complain.client.id === user.id)
+            return complain;
+
+        else {
+          throw new UnauthorizedException();
+        }
+        
+        
+        
        }
        
        
-       async addComplain(complainData: AddComplainDto) {
+       async addComplain(complainData: AddComplainDto, user): Promise<complainEntity>{
+        if (user.role === UserRoleEnum.CLIENT )
+        {
+          const newComplain = this.complainRepository.create(complainData);
+          newComplain.client=user;
+          return await this.complainRepository.save(newComplain);
+        }
+        else {
+          throw new UnauthorizedException();
+        }
+        
+        
+        
+         }
+       
+       
+         async deleteComplain(id: string,user): Promise<unknown> {
           
-           return await this.complainRepository.save(complainData);
-         }
-       
-       
-         async deleteComplain(id: string): Promise<unknown> {
-           const deletedComplain = await this.complainRepository.delete(id);
-           if(! deletedComplain) {
-             throw new NotFoundException(`complain d'id ${id} n'existe pas`);
-           } else {
-             return deletedComplain;
-           }
+          if (user.role === UserRoleEnum.CLIENT ){
+
+            const deletedComplain = await this.complainRepository.delete(id);
+            if(! deletedComplain) {
+              throw new NotFoundException(`complain d'id ${id} n'existe pas`);
+            } else {
+              return deletedComplain;
+            }
+          }
+          else {
+            throw new UnauthorizedException();
+          }
+         
+         
+         
+         
+         
+         
          }
            
            
-         async putClient(id: string, newComplain: updateComplainDto): Promise<complainEntity> {
+         async putComplain(id: string, newComplain: updateComplainDto,user): Promise<complainEntity> {
            const updatedComplain = await this.complainRepository.preload({
              id,
              ...newComplain
-         });
+          });
            console.log('Valeur de retour de preload : ', updatedComplain);
          if (! updatedComplain) {
            throw new NotFoundException(`Le complain d'id ${id} n'existe pas`);
-         } else {
-           return await this.complainRepository.save(updatedComplain);
+         } 
+         if (user.role === UserRoleEnum.ADMINACHAT || updatedComplain.client.id === user.id){
+            return await this.complainRepository.save(updatedComplain);
          }
+         else 
+           throw new UnauthorizedException();
          }
+
+         
 }
+
